chessboard: Create Network Connections Based on Chess Moves

Provides functions to work with directed (asymmetric) and undirected (symmetric) spatial networks. It makes the creation of connectivity matrices easier, i.e. a binary matrix of dimension n x n, where n is the number of nodes (sampling units) indicating the presence (1) or the absence (0) of an edge (link) between pairs of nodes. Different network objects can be produced by 'chessboard': node list, neighbor list, edge list, connectivity matrix. It can also produce objects that will be used later in Moran's Eigenvector Maps (Dray et al. (2006) doi:10.1016/j.ecolmodel.2006.02.015 ) and Asymetric Eigenvector Maps (Blanchet et al. (2008) doi:10.1016/j.ecolmodel.2008.04.001 ), methods available in the package 'adespatial' (Dray et al. (2023) https://CRAN.R-project.org/package=adespatial). This work is part of the FRB-CESAB working group Bridge https://www.fondationbiodiversite.fr/en/the-frb-in-action/programs-and-projects/le-cesab/bridge/.

See also

Author

Maintainer: Nicolas Casajus nicolas.casajus@fondationbiodiversite.fr (ORCID) [copyright holder]

Authors:

Erica Rievrs Borges erica.rievrs@fondationbiodiversite.fr (ORCID)

Eric Tabacchi eric.tabacchi@univ-tlse3.fr (ORCID)

Guillaume Fried guillaume.fried@anses.fr (ORCID)

Nicolas Mouquet nicolas.mouquet@cnrs.fr (ORCID)