Compute the distance between two nodes along a linear shape

Computes distance between two nodes (sites) along a linear shape (e.g. a river, a road, etc.). This function uses parallel computing to speed up the distance computation.

distance_along(
  sites,
  along,
  density = 0.01,
  type = "regular",
  mc.cores = parallel::detectCores() - 1,
  ...
)

Arguments

sites

an sf object of type POINT. A spatial object containing coordinates of sites. Note that the first column must be the site label.

along

an sf object of type LINESTRING. A spatial object containing coordinates of the linear shape (e.g. a river, a road, etc.) to follow while computing distances between two nodes (sites).

density

a numeric of length 1. The density of points to sample on the linear structure. See sf::st_line_sample() for further detail. Default is 1/100.

type

a character of length 1. The method to sample points on the linear shape. Either regular (default) or random.

mc.cores

an integer of length 1. The number of cores to use (must be lesser than the number of cores available on the machine obtained with parallel::detectCores()).

...

other argument to pass to sf::st_line_sample().

Value

A three-column data.frame with:

from, the first node

to, the second node

weight, the distance between the two nodes along the linear shape

Examples

# Import Adour river (LINESTRING) ----
path_to_file <- system.file("extdata", "adour_lambert93.gpkg", 
                            package = "chessboard")
adour_river <- sf::st_read(path_to_file)
#> Reading layer `adour_lambert93' from data source 
#>   `/home/runner/work/_temp/Library/chessboard/extdata/adour_lambert93.gpkg' 
#>   using driver `GPKG'
#> Simple feature collection with 1 feature and 1 field
#> Geometry type: LINESTRING
#> Dimension:     XY
#> Bounding box:  xmin: 334324.8 ymin: 6207072 xmax: 480886.8 ymax: 6308530
#> Projected CRS: RGF93 v1 / Lambert-93

# Import Adour sites ----
path_to_file <- system.file("extdata", "adour_sites_coords.csv", 
                            package = "chessboard")
adour_sites <- read.csv(path_to_file)

# Convert sites to sf object (POINTS) ----
adour_sites <- sf::st_as_sf(adour_sites, coords = 2:3, crs = "epsg:2154")

if (FALSE) {
# Compute distances between pairs of sites along the Adour river ----
dist_along_river <- distance_along(adour_sites, adour_river)
}