Compute the Euclidean distance between two nodes

Computes the Euclidean distance between two nodes using the function sf::st_distance(). If the CRS is not a Cartesian system, the Great Circle distance will be used instead.

distance_euclidean(sites, ...)

Arguments

sites

an sf object of type POINT. A spatial object containing coordinates of sites. Note that the first column must be the site label.

...

other argument to pass to sf::st_distance().

Value

A three-column data.frame with:

from, the first node

to, the second node

weight, the Euclidean distance between the two nodes

Examples

# Import Adour sites ----
path_to_file <- system.file("extdata", "adour_sites_coords.csv", 
                            package = "chessboard")
adour_sites <- read.csv(path_to_file)

# Convert sites to sf object (POINTS) ----
adour_sites <- sf::st_as_sf(adour_sites, coords = 2:3, crs = "epsg:2154")

# Compute distances between pairs of sites along the Adour river ----
distance_euclidean(adour_sites)
#>     from   to     weight
#> 1   S-01 S-01      0.000
#> 2   S-01 S-02  12278.287
#> 3   S-01 S-03  23919.630
#> 4   S-01 S-04  33207.477
#> 5   S-01 S-05  47774.674
#> 6   S-01 S-06  65640.765
#> 7   S-01 S-07  76743.685
#> 8   S-01 S-08  90280.917
#> 9   S-01 S-09 101265.805
#> 10  S-01 S-10 124742.692
#> 11  S-02 S-01  12278.287
#> 12  S-02 S-02      0.000
#> 13  S-02 S-03  11735.397
#> 14  S-02 S-04  21298.857
#> 15  S-02 S-05  35802.527
#> 16  S-02 S-06  53533.547
#> 17  S-02 S-07  64465.474
#> 18  S-02 S-08  78043.772
#> 19  S-02 S-09  89278.651
#> 20  S-02 S-10 114045.217
#> 21  S-03 S-01  23919.630
#> 22  S-03 S-02  11735.397
#> 23  S-03 S-03      0.000
#> 24  S-03 S-04   9786.038
#> 25  S-03 S-05  24125.688
#> 26  S-03 S-06  41798.645
#> 27  S-03 S-07  52963.814
#> 28  S-03 S-08  66789.510
#> 29  S-03 S-09  78653.676
#> 30  S-03 S-10 105473.228
#> 31  S-04 S-01  33207.477
#> 32  S-04 S-02  21298.857
#> 33  S-04 S-03   9786.038
#> 34  S-04 S-04      0.000
#> 35  S-04 S-05  14567.619
#> 36  S-04 S-06  32501.885
#> 37  S-04 S-07  44678.833
#> 38  S-04 S-08  58974.417
#> 39  S-04 S-09  71864.676
#> 40  S-04 S-10 101213.038
#> 41  S-05 S-01  47774.674
#> 42  S-05 S-02  35802.527
#> 43  S-05 S-03  24125.688
#> 44  S-05 S-04  14567.619
#> 45  S-05 S-05      0.000
#> 46  S-05 S-06  18023.365
#> 47  S-05 S-07  31308.309
#> 48  S-05 S-08  46018.955
#> 49  S-05 S-09  60142.113
#> 50  S-05 S-10  92665.191
#> 51  S-06 S-01  65640.765
#> 52  S-06 S-02  53533.547
#> 53  S-06 S-03  41798.645
#> 54  S-06 S-04  32501.885
#> 55  S-06 S-05  18023.365
#> 56  S-06 S-06      0.000
#> 57  S-06 S-07  15572.540
#> 58  S-06 S-08  30181.275
#> 59  S-06 S-09  45826.852
#> 60  S-06 S-10  82008.648
#> 61  S-07 S-01  76743.685
#> 62  S-07 S-02  64465.474
#> 63  S-07 S-03  52963.814
#> 64  S-07 S-04  44678.833
#> 65  S-07 S-05  31308.309
#> 66  S-07 S-06  15572.540
#> 67  S-07 S-07      0.000
#> 68  S-07 S-08  14834.325
#> 69  S-07 S-09  30259.120
#> 70  S-07 S-10  67179.490
#> 71  S-08 S-01  90280.917
#> 72  S-08 S-02  78043.772
#> 73  S-08 S-03  66789.510
#> 74  S-08 S-04  58974.417
#> 75  S-08 S-05  46018.955
#> 76  S-08 S-06  30181.275
#> 77  S-08 S-07  14834.325
#> 78  S-08 S-08      0.000
#> 79  S-08 S-09  16477.810
#> 80  S-08 S-10  55447.537
#> 81  S-09 S-01 101265.805
#> 82  S-09 S-02  89278.651
#> 83  S-09 S-03  78653.676
#> 84  S-09 S-04  71864.676
#> 85  S-09 S-05  60142.113
#> 86  S-09 S-06  45826.852
#> 87  S-09 S-07  30259.120
#> 88  S-09 S-08  16477.810
#> 89  S-09 S-09      0.000
#> 90  S-09 S-10  39509.300
#> 91  S-10 S-01 124742.692
#> 92  S-10 S-02 114045.217
#> 93  S-10 S-03 105473.228
#> 94  S-10 S-04 101213.038
#> 95  S-10 S-05  92665.191
#> 96  S-10 S-06  82008.648
#> 97  S-10 S-07  67179.490
#> 98  S-10 S-08  55447.537
#> 99  S-10 S-09  39509.300
#> 100 S-10 S-10      0.000