Find edges between nodes based on a degree of neighborhood along a linear shape

Finds edges (links) between nodes (sites) based on a degree of neighborhood and along a linear shape (e.g. river or road). The nodes labels (argument nodes) will be ordered to find edges in a directional way (from upstream to downstream along a linear shape).

Note that the detection of edges is only based on the nodes labels (no explicit spatial detection). For instance, if nodes are labelled as S-01, S-02, ..., S-10, the most upstream site will be S-01 and the most downstream S-10.

With a degree of neighborhood of 1, a node will be linked to the first previous node and also to the first next node (undirected network). So, the same edges will be detected twice. If directed = FALSE, this node will be only linked to first next node (directed network from upstream to downstream).

edges_list(nodes, degree = 1, self = FALSE, all = FALSE, directed = FALSE)

Arguments

nodes

a character vector of nodes (sites) labels.

degree

an integer of length 1. The number of neighbors used to define edges.

self

a logical of length 1. If TRUE, a node can be linked to itself. Default is FALSE.

all

a logical of length 1. If TRUE, the missing edges are also returned. Default is FALSE.

directed

a logical of length 1. If FALSE (default), symmetrical edges (e.g. S01-S02 and S02-S01) are returned (undirected network). Otherwise (directed network) only the first edge (e.g. S01-S02) is returned (according to direction of the nodes labels).

Value

A data.frame with four columns:

edge_id: label of the edge

edge: 0 (no edge) or 1 (edge)

from: label of one of the two nodes of the edge

to: label of the other node of the edge

Examples

if (FALSE) {
# Import Adour sites ----
path_to_file <- system.file("extdata", "adour_sites_coords.csv", 
                            package = "chessboard")
adour_sites  <- read.csv(path_to_file)
adour_sites

# Get nodes ----
adour_nodes <- nodes_list(adour_sites$"site")
adour_nodes

# Find edges with 1 degree of neighborhood (undirected network) ----
edges_list(adour_nodes)

# Find edges with 1 degree of neighborhood (directed network) ----
edges_list(adour_nodes, directed = TRUE)
}