Create an edges weights vector

Creates an edges weights vector that can be used in adespatial::aem(). Resulting edges weights equal to 0 will be replaced by 4 x max(w), where max(w) is the maximal weight in the matrix.

edges_weights_vector(x, y)

Arguments

x

a list of length 2. The nodes_by_edges matrix returned by nodes_by_edges_matrix() or adespatial::aem.build.binary().

y

a data.frame with the following three columns: from (the first node of the edge), to (the second node of the edge), and weight (the weight of the edge between the two nodes, e.g. a distance).

Value

An edges weights vector.

Examples

# Import Adour sites ----
path_to_file <- system.file("extdata", "adour_survey_sampling.csv", 
                            package = "chessboard")
adour_sites <- read.csv(path_to_file)

# Select the 15 first sites ----
adour_sites <- adour_sites[1:15, ]

# Create nodes labels ----
adour_sites <- create_nodes_labels(adour_sites, 
                                   location = "location", 
                                   transect = "transect", 
                                   quadrat  = "quadrat")

# Convert sites to sf object (POINTS) ----
adour_sites_sf <- sf::st_as_sf(adour_sites, 
                               coords = c("longitude", "latitude"),
                               crs = "epsg:2154")

# Create edges based on the pawn move (directed network) ----
adour_edges <- create_edges_list(adour_sites, method = "pawn", 
                                 directed = TRUE)

# Create nodes-by-edges matrix ----
adour_matrix <- nodes_by_edges_matrix(adour_edges)

# Compute Euclidean distances between pairs of sites ----
adour_dists <- distance_euclidean(adour_sites_sf)

# Create Edges weights vector ----
edges_weights_vector(adour_matrix, adour_dists)
#>      E-01      E-02      E-03      E-04      E-05      E-06      E-07      E-08 
#> 10001.876 10001.876 10001.876  2500.469  2500.469  2500.469  2500.469  2500.469 
#>      E-09      E-10      E-11      E-12      E-13      E-14      E-15 
#>  2500.469  2500.469  2500.469  2500.469  2500.469  2500.469  2500.469